+++ test/tools/logger.test.js
@@ -8,10 +8,20 @@ describe('logger', () => {
     log.warning('alert')
     expect(console.warn).toHaveBeenCalled()
   })
+  test("Warning with valid string, and verbose false doesn't log a warning", () => {
+    console.warn = jest.fn()
+    log.warning('alert', false)
+    expect(console.warn).not.toHaveBeenCalled()
+  })
   test('Info with valid string, log info', () => {
     console.info = jest.fn()
     log.info('alert')
     expect(console.info).toHaveBeenCalled()
+  })
+  test("Info with valid string, and verbose false doesn't log an info", () => {
+    console.info = jest.fn()
+    log.info('alert', false)
+    expect(console.warn).not.toHaveBeenCalled()
   })
   test('Error with valid string, log error', () => {
     console.error = jest.fn()
@@ -22,5 +32,10 @@ describe('logger', () => {
     console.log = jest.fn()
     log.done('alert')
     expect(console.log).toHaveBeenCalled()
+  })
+  test("Done with valid string, and verbose false doesn't log a log", () => {
+    console.log = jest.fn()
+    log.done('alert', false)
+    expect(console.warn).not.toHaveBeenCalled()
   })
 })