+++ changelog.md
@@ -2,6 +2,11 @@
 
 ## [1.1.0]
 
+### Added
+
+- Add verbose mode. false by default.
+
 ### Changed
 
 - Improved logs message.
+- Improve doc.